requirements.txt
# Langchain Core
Langchain
langchain-Core
#OpenAi Integration
langchain-openai
openai
#Anthropic integration
langchain-anthropic
# google gemini integration
langchain-google
google-generativeai
# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning and Data Science
numpy
pandas
scikit-learn
scipy
matplotlib
seaborn